Update for issue 31
This should fix database connectivity issues. Transactions are now atomic, meaning that a new connection is opened for each query and subsequently closed.

In addition to this a BigNumber polyfill has been added to all code that uses it as teh latest update to the library is breaking (a number of function names have changed).

Finally, a minor update to affiliate number formatting on the account page has been made (I thought this was done previously).

https://github.com/CoinRoster/slotmachine/issues/31