Fix for issue 30
The updates, as listed in the issue description (https://github.com/CoinRoster/slotmachine/issues/30), have been implemented.

Tested with reset jackpot and verified that:

 - jackpot value is coming fro database field directly on both main game and stats page
 - jackpot is being correctly incremented on every spin
 - jackpot is being correctly reset from new *btc_base* field
 - jackpot is correctly credited to winning account / player
 - jackpot is correctly subtracted from "bankroll" investment (the *btc_gains* column is debited with the approrpiate amount, effectively registering as a large loss)

In addition to this, a .gitignore file has been added to exclude third-party libraries that may be installed in live locations (e.g. phpMyAdmin and npm-managed modules)