Fix for issue 39
https://github.com/CoinRoster/slotmachine/issues/39

There were a couple of places in the dividend payments loop that were resetting the main loop counter which caused it to loop infinitely. This section has been updated and after a number of tests appears to be functioning correcly. There should now only be one nightly didivdend payment notification per account along with only one entry on the database.