Minor update for issue 44
Some code was left in the Admin Plugin that was adding an extra 10 Bitcoin to witdrawals (for over-limit testing). Additionally, a miner's fee of 1 Satoshi (0.00000001 Bitcoin) was added to transactions as 0-fee transactions are currently being rejected by the BlockCypher API.

Note that there appears to be some latency between blockchain.info's API and BlockCypher's so that posted transactions may not immediately disappear from a refreshed list (we will probably need to experiment with this to get a good balance).